separate map banner for detail page from global PageBanner component
