+++ src/app/global/components/page-banner/page-banner.vue
@@ -1,31 +1,6 @@
 <template>
   <div class="page-banner">
-    <template v-if="map">
-      <div class="map-wrapper">
-        <template v-if="startingBounds">
-          <nwi-map
-            height="400"
-            :detail-reach-id="Number(reachId)"
-            :include-legend="false"
-            :has-controls="false"
-            :source-layers="$options.sourceLayers"
-            :starting-bounds="startingBounds"
-            fullscreen-target="map-wrapper"
-            hide-result-counter
-          />
-        </template>
-        <template v-else>
-          <utility-block
-            text="Reach Map Unavailable"
-            state="content"
-            theme="dark"
-          />
-        </template>
-        <slot />
-      </div>
-    </template>
-
-    <header v-if="!map">
+    <header>
       <div>
         <h4>{{ subtitle }}</h4>
         <h1>{{ title }}</h1>
@@ -37,16 +12,8 @@
   </div>
 </template>
 <script>
-import bbox from '@turf/bbox'
-import { lineString } from '@turf/helpers'
-import NwiMap from '@/app/views/river-index/components/nwi-map.vue'
-import UtilityBlock from '@/app/global/components/utility-block/utility-block.vue'
 export default {
   name: 'page-banner',
-  components: {
-    'nwi-map': NwiMap,
-    UtilityBlock
-  },
   props: {
     title: {
       type: String,
@@ -56,32 +23,9 @@ export default {
       type: String,
       required: false
     },
-    map: {
-      type: Boolean,
-      required: false
-    },
-    geom: {
-      type: String,
-      required: false
-    },
-    reachId: {
-      type: String,
-      required: false
-    },
     loading: {
       type: Boolean,
       required: false
-    }
-  },
-  sourceLayers: ['reach-segments', 'access'],
-  computed: {
-    startingBounds () {
-      // TODO: get graphql API to return a linestring or geojson instead of this text
-      if (this.geom) {
-        const bounds = this.geom.split(',').map(d => d.split(' '))
-        return bbox(lineString(bounds))
-      }
-      return null
     }
   }
 }
